Move email_accounts and email_groups APIs under 'destinations' namespace
