+++ server/routes/destinations.js
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-export default function(server, services) {
+export default function (server, services) {
   const { destinationsService } = services;
 
   server.route({
@@ -44,5 +44,65 @@ export default function(server, services) {
     path: '/api/alerting/destinations/{destinationId}',
     method: 'DELETE',
     handler: destinationsService.deleteDestination,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_accounts',
+    method: 'GET',
+    handler: destinationsService.getEmailAccounts,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_accounts',
+    method: 'POST',
+    handler: destinationsService.createEmailAccount,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_accounts/{id}',
+    method: 'GET',
+    handler: destinationsService.getEmailAccount,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_accounts/{id}',
+    method: 'PUT',
+    handler: destinationsService.updateEmailAccount,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_accounts/{id}',
+    method: 'DELETE',
+    handler: destinationsService.deleteEmailAccount,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_groups',
+    method: 'GET',
+    handler: destinationsService.getEmailGroups,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_groups',
+    method: 'POST',
+    handler: destinationsService.createEmailGroup,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_groups/{id}',
+    method: 'GET',
+    handler: destinationsService.getEmailGroup,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_groups/{id}',
+    method: 'PUT',
+    handler: destinationsService.updateEmailGroup,
+  });
+
+  server.route({
+    path: '/api/alerting/destinations/email_groups/{id}',
+    method: 'DELETE',
+    handler: destinationsService.deleteEmailGroup,
   });
 }
