--- server/routes/emailGroups.js
@@ -1,48 +1,0 @@
-/*
- *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License").
- *   You may not use this file except in compliance with the License.
- *   A copy of the License is located at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   or in the "license" file accompanying this file. This file is distributed
- *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- *   express or implied. See the License for the specific language governing
- *   permissions and limitations under the License.
- */
-
-export default function(server, services) {
-  const { emailGroupService } = services;
-
-  server.route({
-    path: '/api/alerting/email_groups',
-    method: 'GET',
-    handler: emailGroupService.getEmailGroups,
-  });
-
-  server.route({
-    path: '/api/alerting/email_groups',
-    method: 'POST',
-    handler: emailGroupService.createEmailGroup,
-  });
-
-  server.route({
-    path: '/api/alerting/email_groups/{id}',
-    method: 'GET',
-    handler: emailGroupService.getEmailGroup,
-  });
-
-  server.route({
-    path: '/api/alerting/email_groups/{id}',
-    method: 'PUT',
-    handler: emailGroupService.updateEmailGroup,
-  });
-
-  server.route({
-    path: '/api/alerting/email_groups/{id}',
-    method: 'DELETE',
-    handler: emailGroupService.deleteEmailGroup,
-  });
-}
