+++ server/services/DestinationsService.js
@@ -13,6 +13,8 @@
  *   permissions and limitations under the License.
  */
 
+import _ from 'lodash';
+
 import { CLUSTER } from './utils/constants';
 import { INDEX } from '../../utils/constants';
 
@@ -111,10 +113,7 @@ export default class DestinationsService {
         query_string: {
           fields: ['destination.name', 'destination.type'],
           default_operator: 'AND',
-          query: `*${search
-            .trim()
-            .split(' ')
-            .join('* *')}*`,
+          query: `*${search.trim().split(' ').join('* *')}*`,
         },
       });
     }
@@ -149,7 +148,7 @@ export default class DestinationsService {
     try {
       const resp = await callWithRequest(req, 'search', params);
       const totalDestinations = resp.hits.total.value;
-      const destinations = resp.hits.hits.map(hit => {
+      const destinations = resp.hits.hits.map((hit) => {
         const {
           _source: destination,
           _id: id,
@@ -161,6 +160,268 @@ export default class DestinationsService {
       });
       return { ok: true, destinations, totalDestinations };
     } catch (err) {
+      return { ok: false, err: err.message };
+    }
+  };
+
+  /**
+   *  -----------------------------------------------------
+   *  ----------------- Email Account API -----------------
+   *  -----------------------------------------------------
+   */
+
+  createEmailAccount = async (req, h) => {
+    try {
+      const params = { body: JSON.stringify(req.payload) };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const createResponse = await callWithRequest(req, 'alerting.createEmailAccount', params);
+      return { ok: true, resp: createResponse };
+    } catch (err) {
+      console.error('Alerting - DestinationService - createEmailAccount:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  updateEmailAccount = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const { ifSeqNo, ifPrimaryTerm } = req.query;
+      const params = {
+        emailAccountId: id,
+        ifSeqNo,
+        ifPrimaryTerm,
+        body: JSON.stringify(req.payload),
+      };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const updateResponse = await callWithRequest(req, 'alerting.updateEmailAccount', params);
+      const { _id } = updateResponse;
+      return { ok: true, id: _id };
+    } catch (err) {
+      console.error('Alerting - DestinationService - updateEmailAccount:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  deleteEmailAccount = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const params = { emailAccountId: id };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const deleteResponse = await callWithRequest(req, 'alerting.deleteEmailAccount', params);
+      return { ok: deleteResponse.result === 'deleted' };
+    } catch (err) {
+      console.error('Alerting - DestinationService - deleteEmailAccount:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  getEmailAccount = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const params = { emailAccountId: id };
+      const { callWithRequest } = this.esDriver.getCluster(CLUSTER.ALERTING);
+      const getResponse = await callWithRequest(req, 'alerting.getEmailAccount', params);
+      const emailAccount = _.get(getResponse, 'email_account', null);
+      const ifSeqNo = _.get(getResponse, '_seq_no', null);
+      const ifPrimaryTerm = _.get(getResponse, '_primary_term', null);
+      if (emailAccount) {
+        return { ok: true, resp: emailAccount, ifSeqNo, ifPrimaryTerm };
+      } else {
+        return { ok: false };
+      }
+    } catch (err) {
+      console.error('Alerting - DestinationService - getEmailAccount:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  getEmailAccounts = async (req, h) => {
+    try {
+      const {
+        from = 0,
+        size = 20,
+        search = '',
+        sortDirection = 'desc',
+        sortField = 'name',
+      } = req.query;
+
+      let must = { match_all: {} };
+      if (search.trim()) {
+        must = {
+          query_string: {
+            default_field: 'email_account.name',
+            default_operator: 'AND',
+            query: `*${search.trim().split(' ').join('* *')}*`,
+          },
+        };
+      }
+
+      const sortQueryMap = { name: { 'email_account.name.keyword': sortDirection } };
+
+      let sort = [];
+      const sortQuery = sortQueryMap[sortField];
+      if (sortQuery) sort = sortQuery;
+
+      const params = {
+        body: {
+          from,
+          size,
+          sort,
+          query: {
+            bool: {
+              must,
+            },
+          },
+        },
+      };
+
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const getResponse = await callWithRequest(req, 'alerting.getEmailAccounts', params);
+
+      const totalEmailAccounts = _.get(getResponse, 'hits.total.value', 0);
+      const emailAccounts = _.get(getResponse, 'hits.hits', []).map((result) => {
+        const {
+          _id: id,
+          _seq_no: ifSeqNo,
+          _primary_term: ifPrimaryTerm,
+          _source: emailAccount,
+        } = result;
+        return { id, ...emailAccount, ifSeqNo, ifPrimaryTerm };
+      });
+      return { ok: true, emailAccounts, totalEmailAccounts };
+    } catch (err) {
+      console.error('Alerting - DestinationService - getEmailAccounts:', err);
+      return { ok: false, err: err.message };
+    }
+  };
+
+  /**
+   *  -----------------------------------------------------
+   *  ----------------- Email Group API -------------------
+   *  -----------------------------------------------------
+   */
+
+  createEmailGroup = async (req, h) => {
+    try {
+      const params = { body: JSON.stringify(req.payload) };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const createResponse = await callWithRequest(req, 'alerting.createEmailGroup', params);
+      return { ok: true, resp: createResponse };
+    } catch (err) {
+      console.error('Alerting - DestinationService - createEmailGroup:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  updateEmailGroup = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const { ifSeqNo, ifPrimaryTerm } = req.query;
+      const params = {
+        emailGroupId: id,
+        ifSeqNo,
+        ifPrimaryTerm,
+        body: JSON.stringify(req.payload),
+      };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const updateResponse = await callWithRequest(req, 'alerting.updateEmailGroup', params);
+      const { _id } = updateResponse;
+      return { ok: true, id: _id };
+    } catch (err) {
+      console.error('Alerting - DestinationService - updateEmailGroup:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  deleteEmailGroup = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const params = { emailGroupId: id };
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const deleteResponse = await callWithRequest(req, 'alerting.deleteEmailGroup', params);
+      return { ok: deleteResponse.result === 'deleted' };
+    } catch (err) {
+      console.error('Alerting - DestinationService - deleteEmailGroup:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  getEmailGroup = async (req, h) => {
+    try {
+      const { id } = req.params;
+      const params = { emailGroupId: id };
+      const { callWithRequest } = this.esDriver.getCluster(CLUSTER.ALERTING);
+      const getResponse = await callWithRequest(req, 'alerting.getEmailGroup', params);
+      const emailGroup = _.get(getResponse, 'email_group', null);
+      const ifSeqNo = _.get(getResponse, '_seq_no', null);
+      const ifPrimaryTerm = _.get(getResponse, '_primary_term', null);
+      if (emailGroup) {
+        return { ok: true, resp: emailGroup, ifSeqNo, ifPrimaryTerm };
+      } else {
+        return { ok: false };
+      }
+    } catch (err) {
+      console.error('Alerting - DestinationService - getEmailGroup:', err);
+      return { ok: false, resp: err.message };
+    }
+  };
+
+  getEmailGroups = async (req, h) => {
+    try {
+      const {
+        from = 0,
+        size = 20,
+        search = '',
+        sortDirection = 'desc',
+        sortField = 'name',
+      } = req.query;
+
+      let must = { match_all: {} };
+      if (search.trim()) {
+        must = {
+          query_string: {
+            default_field: 'email_group.name',
+            default_operator: 'AND',
+            query: `*${search.trim().split(' ').join('* *')}*`,
+          },
+        };
+      }
+
+      const sortQueryMap = { name: { 'email_group.name.keyword': sortDirection } };
+
+      let sort = [];
+      const sortQuery = sortQueryMap[sortField];
+      if (sortQuery) sort = sortQuery;
+
+      const params = {
+        body: {
+          from,
+          size,
+          sort,
+          query: {
+            bool: {
+              must,
+            },
+          },
+        },
+      };
+
+      const { callWithRequest } = await this.esDriver.getCluster(CLUSTER.ALERTING);
+      const getResponse = await callWithRequest(req, 'alerting.getEmailGroups', params);
+
+      const totalEmailGroups = _.get(getResponse, 'hits.total.value', 0);
+      const emailGroups = _.get(getResponse, 'hits.hits', []).map((result) => {
+        const {
+          _id: id,
+          _seq_no: ifSeqNo,
+          _primary_term: ifPrimaryTerm,
+          _source: emailGroup,
+        } = result;
+        return { id, ...emailGroup, ifSeqNo, ifPrimaryTerm };
+      });
+      return { ok: true, emailGroups, totalEmailGroups };
+    } catch (err) {
+      console.error('Alerting - DestinationService - getEmailGroups:', err);
       return { ok: false, err: err.message };
     }
   };
