+++ server/services/index.js
@@ -18,8 +18,6 @@ import DestinationsService from './DestinationsService';
 import ElasticsearchService from './ElasticsearchService';
 import MonitorService from './MonitorService';
 import AnomalyDetectorService from './AnomalyDetectorService';
-import EmailAccountService from './EmailAccountService';
-import EmailGroupService from './EmailGroupService';
 
 export {
   AlertService,
@@ -27,6 +25,4 @@ export {
   ElasticsearchService,
   MonitorService,
   AnomalyDetectorService,
-  EmailAccountService,
-  EmailGroupService,
 };
