+++ server/services/utils/constants.js
@@ -17,8 +17,8 @@ export const API_ROUTE_PREFIX = '/_opendistro/_alerting';
 export const MONITOR_BASE_API = `${API_ROUTE_PREFIX}/monitors`;
 export const AD_BASE_API = `/_opendistro/_anomaly_detection/detectors`;
 export const DESTINATION_BASE_API = `${API_ROUTE_PREFIX}/destinations`;
-export const EMAIL_ACCOUNT_BASE_API = `${API_ROUTE_PREFIX}/email_accounts`;
-export const EMAIL_GROUP_BASE_API = `${API_ROUTE_PREFIX}/email_groups`;
+export const EMAIL_ACCOUNT_BASE_API = `${DESTINATION_BASE_API}/email_accounts`;
+export const EMAIL_GROUP_BASE_API = `${DESTINATION_BASE_API}/email_groups`;
 export const DEFAULT_HEADERS = {
   'Content-Type': 'application/json',
   Accept: 'application/json',
